Add third order model and clean UI
Added a third order model function to the AutoFpmMatch and added infrastructure to enable selection of a model type as a parameter in the UI. Cleaned up the UI and removed most global variables for better solutions. Tested and it works, butdoes not seem to give the desired matching for gas phase. Need to also add modelthat matches purely by response factor adjacency.