Add functionality to label duplicate matched compounds as isomer of peak with largest area
Added a function (and manual script duplicateMatch.py) that reads a DataFrame of matched FIDpMS data and handles cases where there are several rows with the same compound name. The basic logic involves treating the peak with largest area as the primary compound and relabeling every other duplicate as an isomer of that compound. In most cases there is no further alteration, but in the case that the primary compound is linear the compound type abbreviation for duplicates is changed to 'B' for Branched Alkanes. There is no way through the UI to disable this feature.