Release alpha deployment
This release includes the first fully-functioning distribution of AutoQuant as a self-standing application. The files necessary to run AutoQuant are found under the new dist/ directory. This includes all python packages and dependencies formatted as-required. To install AutoQuant, simply download the dist/ directory and place the contained AutoQuant directory to your preferred location. From there, edit the properties.json file found under the _internal/ directory and add the preferred AutoQuant files folder (recommended directory is Users/*User*/Documents/AutoQuant). Into this folder you must place the updated resources, response-factors, and images folders from the git repository. To use, simple add the necessary data files according to file structure to the data directory in the files folder. Thanks to everyone who has helped with the project and who will be using it in the future!