Rework QuantUI.py to no longer use subprocesses and instead import external code as packages
Modified AutoFpmMatch.py and AutoQuantification.py such that all important code is contained within one main function that is called by QuantUI.py. This is a departure from the old system, in which subprocesses were created and run every time matching or quantification were performed. This added unnecessary complication and somehow caused a separate Python environment to be run in the terminal rather than the specific subprocess, essentially halting the code's progress. It is unclear what exactly caused this bug to appear, but this new system of importing packages should be simpler and less buggy. It seems to also speed up the execution of eeither matching or quantification, which is interesting considering a subprocess should ideally be faster. Since functions are imported directly, added some clarifying print statements to critical elements of code to indicate that certain segments are running or have run.